+++ examples/features/src/lib.rs
@@ -87,6 +87,7 @@ fn all_tests() -> Vec<wgpu_test::GpuTestInitializer> {
         test_list.push(big_compute_buffers::tests::TWO_BUFFERS);
         test_list.push(cooperative_matrix::tests::COOPERATIVE_MATRIX_F32);
         test_list.push(cooperative_matrix::tests::COOPERATIVE_MATRIX_F16);
+        test_list.push(cooperative_matrix::tests::COOPERATIVE_MATRIX_RECTANGULAR);
     }
 
     test_list